requirements.txt
fastapi==0.110.0
uvicorn==0.27.0

pdfplumber==0.10.2
paddlepaddle==2.6.2
paddleocr==2.7.0.3
drain3==0.9.0
sentence-transformers==2.2.2
FlagEmbedding>=1.2.0
requests==2.31.0
python-dotenv==1.0.1

haystack-ai>=2.4.0,<2.5.0
autogen>=0.2.16,<0.4.0
transformers==4.39.3
torch==2.2.0
faiss-cpu
numpy==1.26.4
scikit-learn==1.4.0
networkx==3.2.1
pyautogen==0.2.30
httpx

torch==2.5.0 torchvision==0.20.0 torchaudio==2.5.0 --index-url https://download.pytorch.org/whl/cpu
transformers==4.39.3
sentence-transformers==2.5.1
tokenizers==0.15.2
uvicorn 
langchain 
langchain-huggingface 